refactor: repackage http request related class
