+++ src/main/java/controller/AbstractController.java
@@ -1,6 +1,6 @@
 package controller;
 
-import http.HttpMethod;
+import http.request.HttpMethod;
 import http.request.HttpRequest;
 import http.response.HttpResponse;
 import http.StatusCode;
