+++ src/test/java/http/request/HttpMethodTest.java
@@ -1,6 +1,7 @@
-package http;
+package http.request;
 
 
+import http.request.HttpMethod;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
