+++ src/test/java/http/request/HttpRequestTest.java
@@ -1,6 +1,5 @@
 package http.request;
 
-import http.HttpMethod;
 import http.Statics;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
