+++ src/test/java/http/request/PathTest.java
@@ -1,5 +1,6 @@
-package http;
+package http.request;
 
+import http.request.Path;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
