+++ src/test/java/http/request/ProtocolTest.java
@@ -1,5 +1,6 @@
-package http;
+package http.request;
 
+import http.request.Protocol;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
