+++ src/test/java/http/request/QueryStringTest.java
@@ -1,5 +1,6 @@
-package http;
+package http.request;
 
+import http.request.QueryString;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
