+++ src/test/java/http/request/RequestHeaderTest.java
@@ -1,5 +1,7 @@
-package http;
+package http.request;
 
+import http.Statics;
+import http.request.RequestHeader;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
