+++ src/test/java/http/request/RequestLineTest.java
@@ -1,5 +1,7 @@
-package http;
+package http.request;
 
+import http.request.HttpMethod;
+import http.request.RequestLine;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
